+++ src/main/java/com/aleksander/restaurant/reservation/dto/ReservationDTO.java
@@ -23,4 +23,6 @@ public class ReservationDTO {
     private LocalDateTime endTime;
 
     private ReservationStatus status;
+
+    private Long parentReservationId;
 }